Rename TransferFunc1DParam to ColorParam
